add primitive premint

+++ NEP5.Contract/Nep5Token.template.cs
@@ -14,6 +14,21 @@ namespace NEP5.Contract
         public static byte Decimals() => D_DECIMALS;
         public static readonly byte[] Owner = "D_OWNER".ToScriptHash();
 
+        //#if D_PREMINT_COUNT > 0
+        //#ifdef D_PREMINT_ADDRESS_0
+        public static readonly byte[] PremintAddress0 = "D_PREMINT_ADDRESS_0".ToScriptHash();
+        public static readonly BigInteger PremintAmount0 = new BigInteger(D_PREMINT_AMOUNT_0);
+        //#endif
+        //#ifdef D_PREMINT_ADDRESS_1
+        public static readonly byte[] PremintAddress1 = "D_PREMINT_ADDRESS_1".ToScriptHash();
+        public static readonly BigInteger PremintAmount1 = new BigInteger(D_PREMINT_AMOUNT_1);
+        //#endif
+        //#ifdef D_PREMINT_ADDRESS_2
+        public static readonly byte[] PremintAddress2 = "D_PREMINT_ADDRESS_2".ToScriptHash();
+        public static readonly BigInteger PremintAmount2 = new BigInteger(D_PREMINT_AMOUNT_2);
+        //#endif
+        //#endif
+        
         public delegate void Action();
         public delegate void Action<in T1>(T1 arg1);
         public delegate void Action<in T1, in T2>(T1 arg1, T2 arg2);
@@ -30,6 +45,9 @@ namespace NEP5.Contract
 
         public static Object Main(string operation, params object[] args)
         {
+            //#if D_PREMINT_COUNT > 0
+            if (operation == Operations.Init) return Init();
+            //#endif
             if (operation == Operations.Owner) return Owner;
             if (operation == Operations.Name) return Name();
             if (operation == Operations.Symbol) return Symbol();
@@ -50,6 +68,24 @@ namespace NEP5.Contract
             return false;
         }
 
+        //#if D_PREMINT_COUNT > 0
+        public static bool Init()
+        {
+            if (Storage.Get(Storage.CurrentContext, Constants.Inited).AsString() == Constants.Inited) return false;
+            //#ifdef D_PREMINT_ADDRESS_0
+            _Mint(PremintAddress0, PremintAmount0);
+            //#endif
+            //#ifdef D_PREMINT_ADDRESS_1
+            _Mint(PremintAddress1, PremintAmount1);
+            //#endif
+            //#ifdef D_PREMINT_ADDRESS_2
+            _Mint(PremintAddress2, PremintAmount2);
+            //#endif
+            Storage.Put(Storage.CurrentContext, Constants.Inited, Constants.Inited);
+            return true;
+        }
+        //#endif
+        
         public static BigInteger BalanceOf(byte[] account)
         {
             return Storage.Get(Storage.CurrentContext, account).AsBigInteger();
@@ -59,9 +95,10 @@ namespace NEP5.Contract
         {
             if (value <= 0) return false;
             if (!Runtime.CheckWitness(from)) return false;
-            if (from == to) return true;
+            if (to.Length != 20) return false;
             BigInteger fromBalance = Storage.Get(Storage.CurrentContext, from).AsBigInteger();
             if (fromBalance < value) return false;
+            if (from == to) return true;
             if (fromBalance == value)
             {
                 Storage.Delete(Storage.CurrentContext, from);
@@ -91,6 +128,7 @@ namespace NEP5.Contract
         public static bool Approve(byte[] originator, byte[] to, BigInteger value)
         {
             if (!Runtime.CheckWitness(originator)) return false;
+            if (to.Length != 20) return false;
             Storage.Put(Storage.CurrentContext, originator.Concat(to), value);
             return true;
         }
@@ -98,6 +136,8 @@ namespace NEP5.Contract
         public static bool TransferFrom(byte[] originator, byte[] from, byte[] to, BigInteger value)
         {
             if (!Runtime.CheckWitness(originator)) return false;
+            if (from.Length != 20) return false;
+            if (to.Length != 20) return false;
             byte[] key = from.Concat(originator);
             BigInteger allowed = Allowance(from, originator);
             BigInteger fromBalance = BalanceOf(from);
@@ -131,6 +171,12 @@ namespace NEP5.Contract
         {
             if (!Runtime.CheckWitness(Owner)) return false;
             if (MintingFinished()) return false;
+            return _Mint(to, value);
+        }
+
+        private static bool _Mint(byte[] to, BigInteger value)
+        {
+            if (to.Length != 20) return false;
             if (value <= 0) return false;
             Storage.Put(Storage.CurrentContext, to, BalanceOf(to) + value);
             Storage.Put(Storage.CurrentContext, Constants.TotalSupply, TotalSupply() + value);
